Add files via upload
Date of Update:21/11/2024
The task has been accomplished. We have got the robot to navigate the entire maze, detect the light intensity at each point, log the coordinates then make sure that it completes one entire round of the maze after which it navigates directly to the brightest spot and come to a rest. We have also made sure to add an LED for the additional reward whose task is to light up in bright red color once it reaches the spot with the highest light intensity.